kanban/pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "pkanban"
version = "0.1.0"
description = "CLI client for Kanban boards"
readme = "README.md"
license = {text = "MIT"}
requires-python = ">=3.8"
classifiers = [
    "Development Status :: 3 - Alpha",
    "Environment :: Console",
    "Intended Audience :: End Users/Desktop",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.8",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
]
dependencies = [
    "requests>=2.31.0",
    "pyyaml>=6.0",
]

[project.scripts]
kanban = "kanban.cli:main"

[project.urls]
Homepage = "https://github.com/pearachute/kanban"
Repository = "https://github.com/pearachute/kanban"

[tool.setuptools.packages.find]
where = ["."]
include = ["kanban*"]
